feat: create provider functions for Lumberjack core
